professor/requirements.txt
requests
python-dotenv
anthropic
feedparser
